- Implemented Nuclear facility grade security aka encryption
- New icon

- Background Infrastructure changed to improve reliablity
