+++ assets/styles/styles.css
@@ -1180,9 +1180,6 @@ body[data-theme="dark"] .table-preview th {
 }
 
 /* Add these specific rules for table cells in dark mode */
-body[data-theme="dark"] .table-preview td,
-body[data-theme="dark"] #spreadsheetContainer 
-
 body[data-theme="dark"] .table-preview td.selected-cell,
 body[data-theme="dark"] #spreadsheetContainer td.selected-cell {
     background-color: rgba(26, 115, 232, 0.1) !important;
@@ -1516,4 +1513,87 @@ body[data-theme="dark"] .confirmation-dialog {
 
 .delete-note-icon {
     color: red !important;
+}
+
+/* Image Support Styles */
+.note-image-container {
+    position: relative;
+    display: inline-block;
+    margin: 5px 0;
+    max-width: 100%;
+}
+
+.note-image {
+    max-width: 100%;
+    height: auto;
+    border-radius: 5px;
+    box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+    transition: all 0.3s ease;
+}
+
+.note-image:hover {
+    box-shadow: 0 4px 8px rgba(0,0,0,0.2);
+}
+
+.image-controls {
+    position: absolute;
+    top: 5px;
+    right: 5px;
+    opacity: 0;
+    transition: opacity 0.3s ease;
+    display: flex;
+    gap: 2px;
+}
+
+.note-image-container:hover .image-controls {
+    opacity: 1;
+}
+
+.image-control-btn {
+    background: rgba(0,0,0,0.7);
+    color: white;
+    border: none;
+    border-radius: 3px;
+    padding: 4px 6px;
+    cursor: pointer;
+    font-size: 12px;
+    transition: background 0.2s ease;
+}
+
+.image-control-btn:hover {
+    background: rgba(0,0,0,0.9);
+}
+
+.delete-btn:hover {
+    background: rgba(220,53,69,0.9);
+}
+
+.resize-btn:hover {
+    background: rgba(0,123,255,0.9);
+}
+
+/* Dark mode image support styles */
+body[data-theme="dark"] .note-image {
+    box-shadow: 0 2px 4px rgba(255,255,255,0.1);
+}
+
+body[data-theme="dark"] .note-image:hover {
+    box-shadow: 0 4px 8px rgba(255,255,255,0.2);
+}
+
+body[data-theme="dark"] .image-control-btn {
+    background: rgba(255,255,255,0.2);
+    color: #eaeaea;
+}
+
+body[data-theme="dark"] .image-control-btn:hover {
+    background: rgba(255,255,255,0.3);
+}
+
+body[data-theme="dark"] .delete-btn:hover {
+    background: rgba(220,53,69,0.8);
+}
+
+body[data-theme="dark"] .resize-btn:hover {
+    background: rgba(0,123,255,0.8);
 }
